[sk] Refactored data loader package
