+++ mage_ai/io/redshift.py
@@ -1,4 +1,4 @@
-from mage_ai.data_loader.base import BaseSQL
+from mage_ai.io.base import BaseSQL
 from pandas import DataFrame
 from redshift_connector import connect
